+++ Modules/Ruleset.Firewall/Private/Convert-ArrayToList.ps1
@@ -51,7 +51,7 @@ PS> Convert-ArrayToList @("192.168.1.1", "192.168.2.1", "172.24.33.100")
 [string[]]
 
 .OUTPUTS
-[string] comma separated list
+[string] Comma separated list
 
 .NOTES
 Following modifications by metablaster: